+++ promotion-api-plugin/src/main/java/hudson/plugins/promoted_builds/Promotion.java
@@ -65,6 +65,9 @@ import org.kohsuke.stapler.StaplerResponse;
  */
 public class Promotion extends AbstractBuild<PromotionProcess,Promotion> implements PromotionRun {
 
+    public static final PermissionGroup PERMISSIONS = new PermissionGroup(Promotion.class, Messages._Promotion_Permissions_Title());
+    public static final Permission PROMOTE = new Permission(PERMISSIONS, "Promote", Messages._Promotion_PromotePermission_Description(), Jenkins.ADMINISTER, PermissionScope.RUN);
+
     public Promotion(PromotionProcess job) throws IOException {
         super(job);
     }
@@ -513,9 +516,6 @@ public class Promotion extends AbstractBuild<PromotionProcess,Promotion> impleme
 
     }
 
-    public static final PermissionGroup PERMISSIONS = new PermissionGroup(Promotion.class, Messages._Promotion_Permissions_Title());
-    public static final Permission PROMOTE = new Permission(PERMISSIONS, "Promote", Messages._Promotion_PromotePermission_Description(), Jenkins.ADMINISTER, PermissionScope.RUN);
-
     @Override
     public int hashCode() {
         return this.getId().hashCode();
@@ -563,59 +563,4 @@ public class Promotion extends AbstractBuild<PromotionProcess,Promotion> impleme
     }
 
     private static final Logger LOGGER = Logger.getLogger(Promotion.class.getName());
-
-    /**
-     * Action, which stores promotion parameters.
-     * This class allows defining custom parameters filtering logic, which is
-     * important for versions after the SECURITY-170 fix.
-     * @since TODO
-     */
-    @Restricted(NoExternalUse.class)
-    public static class PromotionParametersAction extends ParametersAction {
-
-        private List<ParameterValue> unfilteredParameters;
-
-        private PromotionParametersAction(List<ParameterValue> params) {
-            // Pass the parameters upstairs
-            super(params);
-            unfilteredParameters = params;
-        }
-
-        @Override
-        public List<ParameterValue> getParameters() {
-            return Collections.unmodifiableList(filter(unfilteredParameters));
-        }
-
-        private List<ParameterValue> filter(List<ParameterValue> params) {
-            // buildToBePromoted::getParameters() invokes the secured method, hence all
-            // parameters from the promoted build are safe.
-            return params;
-        }
-
-        public static PromotionParametersAction buildFor(
-                @Nonnull AbstractBuild<?, ?> buildToBePromoted,
-                @CheckForNull List<ParameterValue> promotionParams) {
-            if (promotionParams == null) {
-                promotionParams = new ArrayList<ParameterValue>();
-            }
-
-            List<ParameterValue> params = new ArrayList<ParameterValue>();
-
-            //Add the target build parameters first, if the same parameter is not being provided by the promotion build
-            List<ParametersAction> parameters = buildToBePromoted.getActions(ParametersAction.class);
-            for (ParametersAction paramAction : parameters) {
-                for (ParameterValue pvalue : paramAction.getParameters()) {
-                    if (!promotionParams.contains(pvalue)) {
-                        params.add(pvalue);
-                    }
-                }
-            }
-
-            //Add all the promotion build parameters
-            params.addAll(promotionParams);
-
-            // Create list of actions to pass to scheduled build
-            return new PromotionParametersAction(params);
-        }
-    }
 }
